# swagger/openapi.yaml
openapi: 3.0.0
info:
  title: API Plataforma de Adoção de Gatos
  version: 1.0.0
  description: >
    API RESTful para gerenciamento de catálogo de animais e processo de triagem de adoção.
    Desenvolvida para a disciplina de Serviços Web — IFSUL Campus Passo Fundo.
  contact:
    name: IFSUL - Serviços Web

servers:
  - url: http://localhost:3000
    description: Servidor local de desenvolvimento

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Insira o token JWT obtido no endpoint POST /api/auth/login

  schemas:
    Gato:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        nome:
          type: string
          example: Mingau
        sexo:
          type: string
          enum: [M, F]
          example: M
        idade:
          type: integer
          description: Idade em meses
          example: 24
        status_fiv_felv:
          type: boolean
          example: false
        status_adocao:
          type: string
          enum: [Disponível, Em Análise, Adotado]
          example: Disponível
        foto_url:
          type: string
          nullable: true
          example: https://example.com/gato.jpg
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    GatoInput:
      type: object
      required: [nome, sexo, idade, status_fiv_felv]
      properties:
        nome:
          type: string
          example: Mingau
        sexo:
          type: string
          enum: [M, F]
          example: M
        idade:
          type: integer
          example: 24
        status_fiv_felv:
          type: boolean
          example: false
        foto_url:
          type: string
          nullable: true
          example: https://example.com/gato.jpg

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
          example: Maria Silva
        email:
          type: string
          format: email
          example: maria@email.com
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    RegisterInput:
      type: object
      required: [nome, email, senha]
      properties:
        nome:
          type: string
          example: Maria Silva
        email:
          type: string
          format: email
          example: maria@email.com
        senha:
          type: string
          minLength: 6
          example: senha123

    LoginInput:
      type: object
      required: [email, senha]
      properties:
        email:
          type: string
          format: email
          example: maria@email.com
        senha:
          type: string
          example: senha123

    LoginResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          type: object
          properties:
            token:
              type: string
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
            user:
              $ref: '#/components/schemas/User'

    PedidoAdocao:
      type: object
      properties:
        id:
          type: string
          format: uuid
        adotante_id:
          type: string
          format: uuid
        gato_id:
          type: string
          format: uuid
        termos_aceitos:
          type: boolean
          example: true
        links_comprovantes:
          type: array
          items:
            type: string
          example:
            - https://exemplo.com/foto-casa.jpg
        status_pedido:
          type: string
          enum: [Pendente, Aprovado, Rejeitado]
          example: Pendente
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        Adotante:
          $ref: '#/components/schemas/User'
        Gato:
          $ref: '#/components/schemas/Gato'

    PedidoAdocaoInput:
      type: object
      required: [gato_id, termos_aceitos]
      properties:
        gato_id:
          type: string
          format: uuid
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        termos_aceitos:
          type: boolean
          example: true
        links_comprovantes:
          type: array
          items:
            type: string
          example:
            - https://exemplo.com/foto-casa.jpg

    PatchStatusInput:
      type: object
      required: [status_pedido]
      properties:
        status_pedido:
          type: string
          enum: [Pendente, Aprovado, Rejeitado]
          example: Aprovado

    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          type: object

    SuccessListResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        total:
          type: integer
          example: 5
        data:
          type: array
          items:
            type: object

    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        statusCode:
          type: integer
          example: 404
        message:
          type: string
          example: Recurso não encontrado

paths:
  /api/gatos:
    get:
      summary: Lista todos os gatos
      tags: [Gatos]
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [Disponível, Em Análise, Adotado]
          description: Filtra gatos pelo status de adoção
      responses:
        '200':
          description: Lista de gatos retornada com sucesso
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessListResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Gato'
              example:
                status: success
                total: 2
                data:
                  - id: a1b2c3d4-e5f6-7890-abcd-ef1234567890
                    nome: Mingau
                    sexo: M
                    idade: 24
                    status_fiv_felv: false
                    status_adocao: Disponível
                    foto_url: null
        '400':
          description: Status de filtro inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Cadastra um novo gato
      tags: [Gatos]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GatoInput'
            example:
              nome: Mingau
              sexo: M
              idade: 18
              status_fiv_felv: false
              foto_url: https://example.com/mingau.jpg
      responses:
        '201':
          description: Gato cadastrado com sucesso
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Gato'
        '400':
          description: Dados inválidos ou campos obrigatórios ausentes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Token não fornecido ou inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/gatos/{id}:
    get:
      summary: Busca um gato pelo ID
      tags: [Gatos]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: UUID do gato
      responses:
        '200':
          description: Gato encontrado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Gato'
        '404':
          description: Gato não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: error
                statusCode: 404
                message: Gato não encontrado
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Atualiza um gato
      tags: [Gatos]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GatoInput'
      responses:
        '200':
          description: Gato atualizado com sucesso
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Gato'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Gato não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Remove um gato
      tags: [Gatos]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Gato removido com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                status: success
                data:
                  message: Gato removido com sucesso
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Gato não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/register:
    post:
      summary: Cadastro de novo adotante
      tags: [Autenticação]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterInput'
            example:
              nome: Maria Silva
              email: maria@email.com
              senha: senha123
      responses:
        '201':
          description: Usuário cadastrado com sucesso
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '400':
          description: Dados inválidos ou incompletos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: E-mail já cadastrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: error
                statusCode: 409
                message: Este e-mail já está cadastrado
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/login:
    post:
      summary: Login do adotante — retorna JWT
      tags: [Autenticação]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
            example:
              email: maria@email.com
              senha: senha123
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Dados incompletos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: error
                statusCode: 401
                message: Credenciais inválidas
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/pedidos-adocao:
    post:
      summary: Cria um pedido de adoção
      tags: [Pedidos de Adoção]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PedidoAdocaoInput'
            example:
              gato_id: a1b2c3d4-e5f6-7890-abcd-ef1234567890
              termos_aceitos: true
              links_comprovantes:
                - https://exemplo.com/foto-casa.jpg
      responses:
        '201':
          description: Pedido criado. Status do gato alterado para Em Análise.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PedidoAdocao'
        '400':
          description: Termos não aceitos ou dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Gato não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Gato não disponível ou pedido duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: Lista todos os pedidos de adoção
      tags: [Pedidos de Adoção]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de pedidos retornada com sucesso
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessListResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/PedidoAdocao'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/pedidos-adocao/meus:
    get:
      summary: Lista os pedidos do usuário autenticado
      tags: [Pedidos de Adoção]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Pedidos do usuário retornados com sucesso
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessListResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/PedidoAdocao'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/pedidos-adocao/{id}/status:
    patch:
      summary: Atualiza o status de um pedido de adoção
      tags: [Pedidos de Adoção]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: UUID do pedido de adoção
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchStatusInput'
            example:
              status_pedido: Aprovado
      responses:
        '200':
          description: >
            Status atualizado. Se Aprovado: gato marcado como Adotado e demais
            pedidos pendentes para o mesmo gato são automaticamente Rejeitados.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PedidoAdocao'
        '400':
          description: Status inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Pedido não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Pedido já foi finalizado (Aprovado ou Rejeitado)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

tags:
  - name: Gatos
    description: Gerenciamento do catálogo de gatos disponíveis para adoção
  - name: Autenticação
    description: Registro e autenticação de adotantes
  - name: Pedidos de Adoção
    description: Criação e triagem de pedidos de adoção